+++ .env_vars
@@ -1,4 +1,4 @@
 
-PYTHONPATH=django_filter_propertyy;tests
-MYPYPATH=django_filter_property
+PYTHONPATH=django_property_filtery;tests
+MYPYPATH=django_property_filter
