+++ .vscode/settings.json
@@ -1,7 +1,7 @@
 {
     "python.pythonPath": ".venv\\Scripts\\python.exe",
     "python.autoComplete.extraPaths": [
-        "${workspaceFolder}\\django_filter_property"
+        "${workspaceFolder}\\django_property_filter"
     ],
 
     "editor.tabSize": 4,
