+++ dev/Linting/RunPylint.sh
@@ -24,7 +24,7 @@ run_pylint () {
 ##### END OF FUNCTION DEFINITIONS #####
 #######################################
 
-MODULE_NAME=django_filter_property
+MODULE_NAME=django_property_filter
 SCRIPT_PATH="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 PROJ_MAIN_DIR=$SCRIPT_PATH/../..
 MODULE_PATH=$PROJ_MAIN_DIR/$MODULE_NAME
