+++ dev/Testing/RunPytest.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-PACKAGE_ROOT=django_filter_property
+PACKAGE_ROOT=django_property_filter
 SCRIPT_PATH="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 PROJ_MAIN_DIR=$SCRIPT_PATH/../..
 pushd "$PROJ_MAIN_DIR"
