+++ django_property_filter/conf.py
