+++ tests/django_property_filter_tests/filters/__init__.py
@@ -1,0 +1,1 @@
+"""Package to support filtering by Property in Django."""
