+++ tests/django_property_filter_tests/filters/test_property_number_filter.py
@@ -1,0 +1,8 @@
+
+
+
+
+# TODO - TEST Number Filter
+
+# TODO - Automate to test all supported filter methods (from conf.py)
+
