+++ tests/django_property_filter_tests/test_common.py
@@ -1,0 +1,16 @@
+
+
+# TODO - Implement Tests for Filterset Generators to ensure the function generator works
+
+
+
+
+
+'''
+def test_get_valid_django_filter
+
+def test_get_valid_django_filter_propert
+'''
+
+
+
